requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
faker
openai
cerebras-cloud-sdk
requests
